+++ catlogger_test.go
@@ -1,17 +1,12 @@
 package catlogger
 
 import "testing"
-
-func assert(t *testing.T, ok bool, message string) {
-	if !ok {
-		t.Error(message)
-	}
-}
+import "github.com/stretchr/testify/assert"
 
 func TestLogger(t *testing.T) {
 	logger := MakeLogger("foo,bar", "hello", false)
-	assert(t, logger != nil, "could not make logger")
-	assert(t, logger.HasCategory("foo"), "logger lacks category foo")
-	assert(t, logger.HasCategory("bar"), "logger lacks category bar")
-	assert(t, !logger.HasCategory("baz"), "logger has category baz")
+	assert.NotNil(t, logger, "could not make logger")
+	assert.True(t, logger.HasCategory("foo"), "logger lacks category foo")
+	assert.True(t, logger.HasCategory("bar"), "logger lacks category bar")
+	assert.True(t, !logger.HasCategory("baz"), "logger has category baz")
 }