+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for `github.com/MikeTaylor/catlogger`
 
+## IN PROGRESS
+
+* Use github.com/stretchr/testify/assert for testing rather than a hand-rolled `assert` function.
+
 ## [0.0.2](https://github.com/MikeTaylor/catlogger/tree/v0.0.2) (2023-11-08)
 
 * Clarify [`Makefile`](Makefile) comments on linting prerequisites.